Refactor scraper into package
